refactor(desktop): move beta terms state to tasks
Generated-By: PostHog Desktop
Task-Id: 0e527ab7-6576-4a11-9224-06c608e31b7d
