+++ products/legal_documents/backend/tests/test_api.py
@@ -20,7 +20,7 @@ from posthog.models.activity_logging.activity_log import ActivityLog
 from posthog.models.organization import Organization, OrganizationMembership
 
 from products.legal_documents.backend.facade import api as legal_api
-from products.legal_documents.backend.models import DesktopBetaTermsAcceptance, LegalDocument
+from products.legal_documents.backend.models import LegalDocument
 
 BAA_PAYLOAD = {
     "document_type": "BAA",
@@ -35,33 +35,6 @@ DPA_PAYLOAD = {
     "company_address": "1 Analytics Way, SF CA",
     "representative_email": "ada@acme.example",
 }
-
-
-class TestDesktopBetaTermsAPI(APIBaseTest):
-    def setUp(self) -> None:
-        super().setUp()
-        self.url = f"/api/organizations/{self.organization.id}/desktop_beta_terms/"
-
-    def test_admin_can_accept_terms_idempotently(self) -> None:
-        self.organization_membership.level = OrganizationMembership.Level.ADMIN
-        self.organization_membership.save()
-
-        first_response = self.client.post(self.url)
-        second_response = self.client.post(self.url)
-
-        self.assertEqual(first_response.status_code, status.HTTP_200_OK)
-        self.assertEqual(second_response.status_code, status.HTTP_200_OK)
-        self.assertEqual(DesktopBetaTermsAcceptance.objects.filter(organization=self.organization).count(), 1)
-        self.assertTrue(self.client.get(self.url).json()["is_desktop_beta_terms_accepted"])
-
-    def test_member_can_check_but_not_accept_terms(self) -> None:
-        check_response = self.client.get(self.url)
-        accept_response = self.client.post(self.url)
-
-        self.assertEqual(check_response.status_code, status.HTTP_200_OK)
-        self.assertFalse(check_response.json()["is_desktop_beta_terms_accepted"])
-        self.assertEqual(accept_response.status_code, status.HTTP_403_FORBIDDEN)
-        self.assertFalse(DesktopBetaTermsAcceptance.objects.exists())
 
 
 def _billing_with_addons(addon_types_subscribed: set[str]) -> dict[str, Any]:
